+++ tasmota/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add commands ``LedPwmOn 0..255``, ``LedPwmOff 0..255`` and ``LedPwmMode1 0/1`` to control led brightness by George (#8491)
 - Add Three Phase Export Active Energy to SDM630 driver
 - Add wildcard pattern ``?`` for JSON matching in rules
+- Add support for unique MQTTClient (and inherited fallback topic) by full Mac address using ``mqttclient DVES_%12X`` (#8300)
 
 ### 8.3.1.1 20200518